requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
streamlit>=1.32.0
Pillow>=9.4.0
numpy>=1.24.0
